--- src/components/src/main/AddToDoDrawer.tsx
@@ -1,43 +1,0 @@
-"use client";
-
-import * as React from "react";
-
-import { Button } from "@/components/ui/button";
-import {
-  Drawer,
-  DrawerClose,
-  DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerTitle,
-  DrawerTrigger,
-} from "@/components/ui/drawer";
-import { Input } from "@/components/ui/input";
-import { Send } from "lucide-react";
-
-export function AddToDoDrawer({ children }: { children: React.ReactNode }) {
-  return (
-    <Drawer>
-      <DrawerTrigger asChild>{children}</DrawerTrigger>
-      <DrawerContent>
-        <div className="mx-auto w-full max-w-sm">
-          <DrawerHeader>
-            <DrawerTitle>{`Please enter Today's goal!`}</DrawerTitle>
-            <DrawerDescription>
-              <Input />
-            </DrawerDescription>
-          </DrawerHeader>
-          <DrawerFooter>
-            <Button>
-              <Send />
-            </Button>
-            <DrawerClose asChild>
-              <Button variant="outline">Cancel</Button>
-            </DrawerClose>
-          </DrawerFooter>
-        </div>
-      </DrawerContent>
-    </Drawer>
-  );
-}
